hotfix/acm-004 Add Exception for Bad Request.

+++ src/main/java/com/spduniversity/controllers/CommentReactionController.java
@@ -2,6 +2,7 @@ package com.spduniversity.controllers;
 
 import com.spduniversity.dto.CommentReactionDto;
 import com.spduniversity.entities.comments.CommentReaction;
+import com.spduniversity.exceptions.comment_reactions.CommentReactionBadRequestException;
 import com.spduniversity.exceptions.comment_reactions.CommentReactionNoContentException;
 import com.spduniversity.services.CommentReactionService;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +19,9 @@ public class CommentReactionController {
     @GetMapping("/comment-reactions/{commentId}/{commentReaction}")
     int getTotalReactionTypeByCommentId(@PathVariable("commentId") int commentId,
                                         @PathVariable("commentReaction") String commentReaction) {
+        if (commentReaction == null) {
+            throw new CommentReactionBadRequestException();
+        }
         return commentReactionService.getTotalReactionTypeByCommentId(commentId, commentReaction);
     }
 
@@ -33,6 +37,9 @@ public class CommentReactionController {
     @DeleteMapping("/comment-reactions/{commentId}/{commentReaction}")
     void deleteCommentReaction(@PathVariable("commentId") int commentId,
                                @PathVariable("commentReaction") String commentReaction) {
+        if (commentReaction == null) {
+            throw new CommentReactionBadRequestException();
+        }
         commentReactionService.deleteLastRecordByCommentId(commentId, commentReaction);
     }
 }
